Refactor hooks and providers for additional simplification
