--- packages/ui-components/src/components/Chat/hooks/index.ts
@@ -1,5 +1,0 @@
-export * from './useFetchNotification';
-export * from './useIsTabActive';
-
-export {default as useFetchNotification} from './useFetchNotification';
-export {default as useUnstoppableMessaging} from './useUnstoppableMessaging';
