+++ packages/ui-components/src/components/Chat/index.ts
@@ -1,5 +1,3 @@
 export * from './SupportBubble';
 export * from './UnstoppableMessaging';
-export * from './hooks';
 export * from './types';
-export {default as UnstoppableMessagingProvider} from './provider/UnstoppableMessagingProvider';
