--- packages/ui-components/src/components/Chat/provider/index.ts
@@ -1,2 +1,0 @@
-export * from './UnstoppableMessagingProvider';
-export {default as UnstoppableMessagingProvider} from './UnstoppableMessagingProvider';
