+++ packages/ui-components/src/hooks/index.ts
@@ -1,4 +1,4 @@
 export * from './useEnsDomainStatus';
-export {default as usePrevious} from './usePrevious';
 export {default as useTokenGallery} from './useTokenGallery';
+export {default as useUnstoppableMessaging} from './useUnstoppableMessaging';
 export {default as useWeb3Context} from './useWeb3Context';
