+++ packages/ui-components/src/providers/UnstoppableMessagingProvider.tsx
@@ -1,8 +1,6 @@
-import {SnackbarProvider} from 'notistack';
 import React, {useState} from 'react';
 
-import {TranslationProvider} from '../../../lib';
-import {Web3ContextProvider} from '../../../providers';
+import BaseProvider from './BaseProvider';
 
 type Props = {
   children: React.ReactNode;
@@ -45,20 +43,11 @@ const UnstoppableMessagingProvider: React.FC<Props> = ({children}) => {
   };
 
   return (
-    <TranslationProvider>
-      <SnackbarProvider
-        anchorOrigin={{
-          vertical: 'bottom',
-          horizontal: 'left',
-        }}
-      >
-        <Web3ContextProvider>
-          <UnstoppableMessagingContext.Provider value={value}>
-            {children}
-          </UnstoppableMessagingContext.Provider>
-        </Web3ContextProvider>
-      </SnackbarProvider>
-    </TranslationProvider>
+    <BaseProvider>
+      <UnstoppableMessagingContext.Provider value={value}>
+        {children}
+      </UnstoppableMessagingContext.Provider>
+    </BaseProvider>
   );
 };
