+++ packages/ui-components/src/providers/index.ts
@@ -1,5 +1,9 @@
+export * from './BaseProvider';
 export * from './TokenGalleryProvider';
+export * from './UnstoppableMessagingProvider';
 export * from './Web3ContextProvider';
 
+export {default as BaseProvider} from './BaseProvider';
 export {default as TokenGalleryProvider} from './TokenGalleryProvider';
+export {default as UnstoppableMessagingProvider} from './UnstoppableMessagingProvider';
 export {default as Web3ContextProvider} from './Web3ContextProvider';
